still working on getting location services
